// src/localization/languages/enum.ts
enum TKeys {
  CLOSE_SEARCH = 'closeSearch',
  STORES = 'stores',
  SEARCH = 'search',
  BUY = 'buy',
  YOUR_LIST = 'yourList',
  AVAILABLE = 'available',
  CLOSED = 'closed',
  WEEK_DAYS_MONDAY = 'weekDays.monday',
  WEEK_DAYS_TUESDAY = 'weekDays.tuesday',
  WEEK_DAYS_WEDNESDAY = 'weekDays.wednesday',
  WEEK_DAYS_THURSDAY = 'weekDays.thursday',
  WEEK_DAYS_FRIDAY = 'weekDays.friday',
  WEEK_DAYS_SATURDAY = 'weekDays.saturday',
  WEEK_DAYS_SUNDAY = 'weekDays.sunday',
  MOST_SEARCHED_PRODUCTS = 'mostSearchedProducts',
  CATEGORIES = 'categories',
  SEARCHING_FOR = 'searchingFor',
  LOAD_MORE_STORE = 'LOAD_MORE_STORE',
  INDICATIONS = 'INDICATIONS',
  CALL_ME = 'CALL_ME',
  SAVED = 'SAVED',
  OPEN = 'OPEN',
  CLOSE = 'CLOSE',
  CLOSES_AT = 'CLOSES_AT',
  DAY_0 = 'DAY_0',
  DAY_1 = 'DAY_1',
  DAY_2 = 'DAY_2',
  DAY_3 = 'DAY_3',
  DAY_4 = 'DAY_4',
  DAY_5 = 'DAY_5',
  DAY_6 = 'DAY_6',
  GOT_IT = 'GOT_IT',
  SCROLL_DOWN_TO_VIEW_MORE_OFFERS = 'SCROLL_DOWN_TO_VIEW_MORE_OFFERS',
  OFFERS_TO_SEE = 'OFFERS_TO_SEE',
  OFFER_TO_SEE = 'OFFER_TO_SEE',
  OPTIONS_AVAILABLE = 'OPTIONS_AVAILABLE',
  OPTIONS = 'OPTIONS',
  AVAILABLE_IN = 'AVAILABLE_IN',
  MULTI_IMAGE = 'MULTI_IMAGE',
  COMBO = 'COMBO',
  DESCRIPTION = 'DESCRIPTION',
  VIEW_MORE = 'VIEW_MORE',
  VIEW_LESS = 'VIEW_LESS',
}

enum LKeys {
  IT = 'it',
  IT_IT = 'it-IT',
  EN = 'en',
  EN_UK = 'en-UK',
  EN_US = 'en-US',
  EN_AU = 'en-AU',
  EN_GB = 'en-GB',
  EN_NZ = 'en-NZ',
  EN_ZA = 'en-ZA',
  ES = 'es',
  ES_ES = 'es-ES',
  ES_AR = 'es-AR',
  ES_CL = 'es-CL',
  ES_CO = 'es-CO',
  ES_EC = 'es-EC',
  ES_MX = 'es-MX',
  ES_PE = 'es-PE',
  ES_US = 'es-US',
  PT = 'pt',
  PT_BR = 'pt-BR',
  PT_PT = 'pt-PT',
  FR = 'fr',
  FR_FR = 'fr-FR',
}

export { TKeys, LKeys }